[specifications] move condition fontforge_check_results into general specificaion
